chore: Move unit tests to the right folder
